feat: add ResizableSettingsPanelLayout for improved dashboard and worksheet artifact UI
