+++ apps/sqlrooms-cli-ui/src/workspace/WorksheetArtifact.tsx
@@ -3,7 +3,6 @@ import {
   BlockDocumentChartRendererProvider,
   BlockDocumentArtifact,
   BlockDocumentStatefulBlockRendererProvider,
-  BlockSettingsPanel,
   type BlockDocumentStatefulBlockRenderer,
   type BlockDocumentStatefulBlockRendererProps,
   type Editor,
@@ -11,14 +10,9 @@ import {
 import type {RoomPanelComponent} from '@sqlrooms/layout';
 import {DataTableBlockRenderer, ChartBlockRenderer} from '@sqlrooms/mosaic';
 import {PythonBlock} from '@sqlrooms/python/block';
-import {
-  ResizableHandle,
-  ResizablePanel,
-  ResizablePanelGroup,
-  ResizablePanelHandle,
-} from '@sqlrooms/ui';
-import {FC, useCallback, useEffect, useMemo, useRef, useState} from 'react';
+import {FC, useCallback, useEffect, useMemo, useState} from 'react';
 import {experimentalEnabled, useRoomStore} from '../store';
+import {ResizableSettingsPanelLayout} from './ResizableSettingsPanelLayout';
 import {
   createStatefulBlockTypes,
   type StatefulBlockArtifactType,
@@ -163,8 +157,6 @@ export const WorksheetArtifact: RoomPanelComponent = ({panelId, meta}) => {
     (state) => state.artifacts.renameArtifact,
   );
   const [editor, setEditor] = useState<Editor | null>(null);
-  const panelRef = useRef<ResizablePanelHandle>(null);
-  const [isOpen, setIsOpen] = useState(true);
 
   useEffect(() => {
     if (artifact?.type === 'worksheet') {
@@ -192,26 +184,6 @@ export const WorksheetArtifact: RoomPanelComponent = ({panelId, meta}) => {
     [artifactId, renameArtifact],
   );
 
-  useEffect(() => {
-    if (isOpen) {
-      panelRef.current?.expand();
-    } else {
-      panelRef.current?.collapse();
-    }
-  }, [isOpen]);
-
-  const onResize = () => {
-    const isCollapsed = panelRef.current?.isCollapsed();
-
-    if (isCollapsed && isOpen) {
-      setIsOpen(false);
-    }
-
-    if (!isCollapsed && !isOpen) {
-      setIsOpen(true);
-    }
-  };
-
   if (!artifact || artifact.type !== 'worksheet') {
     return null;
   }
@@ -222,35 +194,14 @@ export const WorksheetArtifact: RoomPanelComponent = ({panelId, meta}) => {
         renderers={statefulBlockRenderers}
         blockTypes={statefulBlockTypes}
       >
-        <ResizablePanelGroup orientation="horizontal" className="h-full">
-          <ResizablePanel>
-            <BlockDocumentArtifact
-              artifactId={artifactId}
-              title={artifact.title}
-              onTitleChange={handleTitleChange}
-              onEditorReady={setEditor}
-            />
-          </ResizablePanel>
-          <ResizableHandle withHandle />
-          <ResizablePanel
-            ref={panelRef}
-            defaultSize={300}
-            minSize={300}
-            maxSize="35%"
-            className="overflow-hidden"
-            collapsible={true}
-            collapsedSize={10}
-            onResize={onResize}
-          >
-            {isOpen && (
-              <BlockSettingsPanel
-                className="border-l"
-                editor={editor}
-                documentId={artifactId}
-              />
-            )}
-          </ResizablePanel>
-        </ResizablePanelGroup>
+        <ResizableSettingsPanelLayout editor={editor} documentId={artifactId}>
+          <BlockDocumentArtifact
+            artifactId={artifactId}
+            title={artifact.title}
+            onTitleChange={handleTitleChange}
+            onEditorReady={setEditor}
+          />
+        </ResizableSettingsPanelLayout>
       </BlockDocumentStatefulBlockRendererProvider>
     </BlockDocumentChartRendererProvider>
   );
